Revert "Remove the key-converter module and move corresponding classes to the (#796)"
This reverts commit c555ca76840e6738ac35895f54f9cc7a656a6282.
