--- enclave/enclave-api/src/test/resources/META-INF/services/com.quorum.tessera.key.vault.KeyVaultServiceFactory
@@ -1,2 +1,0 @@
-com.quorum.tessera.enclave.MockAzureKeyVaultServiceFactory
-com.quorum.tessera.enclave.MockHashicorpKeyVaultServiceFactory
