+++ key-pair-converter/src/test/java/com/quorum/tessera/keypairconverter/MockHashicorpKeyVaultServiceFactory.java
@@ -1,9 +1,9 @@
-package com.quorum.tessera.enclave;
+package com.quorum.tessera.keypairconverter;
 
 import com.quorum.tessera.config.Config;
 import com.quorum.tessera.config.KeyVaultType;
 import com.quorum.tessera.config.util.EnvironmentVariableProvider;
-import com.quorum.tessera.config.vault.data.AzureGetSecretData;
+import com.quorum.tessera.config.vault.data.HashicorpGetSecretData;
 import com.quorum.tessera.key.vault.KeyVaultService;
 import com.quorum.tessera.key.vault.KeyVaultServiceFactory;
 
@@ -11,12 +11,12 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class MockAzureKeyVaultServiceFactory implements KeyVaultServiceFactory {
+public class MockHashicorpKeyVaultServiceFactory implements KeyVaultServiceFactory {
     @Override
     public KeyVaultService create(Config config, EnvironmentVariableProvider envProvider) {
         KeyVaultService mock = mock(KeyVaultService.class);
 
-        when(mock.getSecret(any(AzureGetSecretData.class)))
+        when(mock.getSecret(any(HashicorpGetSecretData.class)))
             .thenReturn("publicSecret")
             .thenReturn("privSecret");
 
@@ -25,6 +25,6 @@ public class MockAzureKeyVaultServiceFactory implements KeyVaultServiceFactory {
 
     @Override
     public KeyVaultType getType() {
-        return KeyVaultType.AZURE;
+        return KeyVaultType.HASHICORP;
     }
 }
