+++ key-pair-converter/src/test/resources/META-INF/services/com.quorum.tessera.key.vault.KeyVaultServiceFactory
@@ -1,0 +1,2 @@
+com.quorum.tessera.keypairconverter.MockAzureKeyVaultServiceFactory
+com.quorum.tessera.keypairconverter.MockHashicorpKeyVaultServiceFactory
